backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
python-multipart
python-jose[cryptography]
validators==0.22.0
extruct==0.16.0
lxml
httpx==0.25.2
sqlalchemy
psycopg2-binary
passlib[bcrypt]
email-validator
aiohttp==3.9.1
brotli==1.1.0
authlib==1.2.1
itsdangerous==2.1.2

# Development dependencies
black==23.11.0
ruff==0.1.6
mypy==1.7.0
pytest==7.4.3
pytest-asyncio==0.21.1  # For testing async functions 